[FE] fix: 캐러셀 arrow 공통 컴포넌트로 변경
