+++ front/src/pages/CollectionPage/CollectionSmallSet/CollectionSmallSet.js
@@ -4,6 +4,7 @@ import CollectionSmallBooks from './CollectionSmallBooks';
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
+import { NextArrow, PrevArrow } from '../../../components/CarouselArrows';
 
 const SlickSlider = styled.div`
   width: 100%;
@@ -23,41 +24,6 @@ const SlickSlider = styled.div`
     z-index: 100;
   }
 `;
-
-const ArrowContainer = styled.div`
-  width: 25px;
-  height: 25px;
-`;
-const NextArrow = (props) => {
-  const { className, style, onClick } = props;
-  return (
-    <ArrowContainer
-      className={className}
-      style={{ ...style }}
-      onClick={onClick}
-    >
-      <img
-        src={process.env.PUBLIC_URL + '/images/carousel_next_icon.svg'}
-        alt="carousel next"
-      />
-    </ArrowContainer>
-  );
-};
-const PrevArrow = (props) => {
-  const { className, style, onClick } = props;
-  return (
-    <ArrowContainer
-      className={className}
-      style={{ ...style }}
-      onClick={onClick}
-    >
-      <img
-        src={process.env.PUBLIC_URL + '/images/carousel_prev_icon.svg'}
-        alt="carousel next"
-      />
-    </ArrowContainer>
-  );
-};
 
 const CollectionSmallSetContainer = styled.div`
   width: 100%;
